Lightweight Android.bp writer in orchestrator
This adds several python classes for the various components of an
Android.bp file (e.g. comments, soong_modules, properties etc).

Each class accepts a formatter callback that will be used to serialize
its state to a string. Some default formatters are implemented, which
match the output of `bpfmt` for simple cases.

Test: python -m test_build_file_generator

Change-Id: I9f2d5ffd2211b7d98f47a5a1487585a00357e9bf