Support "*" wildcard in API export
The current implementation of ApiPackageFinder matches the top-level
contribution target using the name of the api_domain in
api_packages.json. Since mcombo files accept "*" as a wildcard to match
all apexes, add a new function to the ApiPackageFinder API that accepts
a generic filter.

api_packages.json now accepts an additional key `is_apex` to identify
whether an api_domain is an apex

Test: python -m test_find_api_packages

Change-Id: I0570948e7975a0e949ad5665952ff7a073d48990