Generate Android.bp files for older versions of CC PublicAPI (NDK)
Currently, `out/api_surfaces` contains Android.bp modules for
`version=current`. This causes build errors for importing inner trees
with cc_library(s) that set sdk_version!=current, since those prebuilts
are not present in the assembled api_surfaces directory.

As a short-term hack, unconditionally generate Android.bp files for NDK
(PublicAPI) libraries starting from 1. Rely on ndkstubgen for API
correctness. As an example, if an API was introduced in 29, it will be
annotated as #introduced=29 in foo.map.txt, and therefore not appear in
stubs older than 29.

Bug: 265962882
Test: multitree_build vendor/vendorimage (top of CL stack)
Change-Id: I4a6bc5d5147ee8cab3e41903d2516ce7f4230e05