Make apex builds work
cc_api_library is aggressive about shadowing the implemention library
that already exists in that tree. This was causing apex variants of
libraries to fail (build orchestrator would give it an empty src and
headers).

This CL is not a principled solution, but introduces further hacks
- Create a phony target for src of the top-level cc_api_library per
  arch. This property will be a no-op eventually.
- Add headers to the top-level cc_api_library. This property should also
  be a no-op in the future, but is necessary now to make apex builds
  succeed.

Also created phony targets into system and apex inner trees for
out/api_surfaces stubs. Previously, it was done just for vendor.

Test: multitree_build system/vendorimage
Test: multitree_build system/droid FAILS!
(because it tries to build recovery image using stubs b/255674803)

Change-Id: I9581d7ebc63a19740194daeac9fba98ab62d5471