Delete the api_domain Soong modules
These modules specify the api exports of an api_domain, which is then used
in Multi-tree API export. The modules were handcrafted by going through
the tree and classifying the exports as ABI-stable or not.

Since the original use case is not as relevant in the immediate future,
delete these modules to prevent bit rot.

Test: TH
Bug: 273723791
Change-Id: I90e4acac8847e523142882d2cf5e335bd2d9c6a1